+++ src/main/java/acme/entities/project/UserStory.java
@@ -1,12 +1,11 @@
 
-package acme.entities;
+package acme.entities.project;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.ManyToOne;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
@@ -18,17 +17,12 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class Project extends AbstractEntity {
+public class UserStory extends AbstractEntity {
 
 	/**
 	 * 
 	 */
 	private static final long	serialVersionUID	= 1L;
-
-	@Pattern(regexp = "[A-Z]{3}-[0-9]{4}")
-	@NotBlank
-	@Column(unique = true)
-	String						code;
 
 	@NotBlank
 	@Length(max = 76)
@@ -36,17 +30,20 @@ public class Project extends AbstractEntity {
 
 	@NotBlank
 	@Length(max = 101)
-	@Column(name = "Abstract")
-	String						abstracto;
+	String						description;
 
-	String						fatalError;
 	@NotNull
-	@Min(0)
-	Integer						cost;
+	@Min(1)
+	Integer						estimatedCost;
 
+	@NotBlank
+	@Length(max = 101)
+	String						acceptanceCriteria;
 	@URL
 	String						link;
 
-	Boolean						draftMode;
+	@NotNull
+	@ManyToOne(optional = false)
+	Project						project;
 
 }
